requirements.txt
# cwprep dependencies

# Configuration file support
pyyaml>=6.0
python-dotenv>=1.0
